refactored Landing page
